Split controller.go off of the Service reconciler
